Move NPathComplexity from metrics to design
